+++ lib/FileMoveObject.d.ts
@@ -1,0 +1,6 @@
+declare class FileMoveObject {
+    inputPath: string;
+    relativeOutputPath: string;
+    constructor(inputPath: string, relativeOutputPath: string);
+}
+export default FileMoveObject;