+++ src/FileExtractor.ts
@@ -17,17 +17,15 @@ var notHttpRegExp = /^(?!https?).*/g;
 var isRelativePath = /\.[\/]/;
 
 var processSourceReplacement = function(wholeMatch: string, tag: string, before: string, wrapBefore: string,
-    wrap: string, val: string, after: string, closingSlash: string, files: FileMoveObject[])
-{
+    wrap: string, val: string, after: string, closingSlash: string, files: FileMoveObject[]) {
+
     var ret = wholeMatch;
 
     // update and extract if local file
     if(val.match(notHttpRegExp)) {
         var fileName = path.basename(val);
 
-        var file: FileMoveObject = new FileMoveObject();
-
-        file.inputPath = val;
+        var inputPath = val;
 
         // replace val
         switch(tag.toLowerCase()) {
@@ -37,16 +35,17 @@ var processSourceReplacement = function(wholeMatch: string, tag: string, before:
             case "link": val = `assets/css/${fileName}`; break;
         }
 
-        file.relativeOutputPath = val;
+        var outputPath = val;
 
-        if(file.inputPath !== file.relativeOutputPath)
-            files.push(file);
+        if(inputPath !== outputPath) {
+            files.push(new FileMoveObject(inputPath, outputPath));
+        }
 
         switch(tag.toLowerCase()) {
             case "img":
             case "source":
-            case "script": ret = `<${tag} ${before?before:""}src=${wrap}${val}${wrap}${after?after:""}${closingSlash?closingSlash:""}>`; break;
-            case "link": ret = `<${tag} ${before?before:""}href=${wrap}${val}${wrap}${after?after:""}${closingSlash?closingSlash:""}>`; break;
+            case "script": ret = `<${tag} ${before ? before : ""}src=${wrap}${val}${wrap}${after ? after : ""}${closingSlash ? closingSlash : ""}>`; break;
+            case "link": ret = `<${tag} ${before ? before : ""}href=${wrap}${val}${wrap}${after ? after : ""}${closingSlash ? closingSlash : ""}>`; break;
         }
     }
 
@@ -80,9 +79,9 @@ function copyFile(source: string, target: string, ignoreNotExistent?: boolean) {
             return;
         }
 
-        var ensureDirectoryExistence = function (filePath: string) {
+        var ensureDirectoryExistence = function(filePath: string) {
             var dirname = path.dirname(filePath);
-            if (fs.existsSync(dirname)) {
+            if(fs.existsSync(dirname)) {
                 return true;
             }
             ensureDirectoryExistence(dirname);
@@ -92,14 +91,14 @@ function copyFile(source: string, target: string, ignoreNotExistent?: boolean) {
 
         var cbCalled = false;
         var rd = fs.createReadStream(source);
-        rd.on("error", function (err: any) {
+        rd.on("error", function(err: any) {
             reject(err);
         });
         var wr = fs.createWriteStream(target);
-        wr.on("error", function (err: any) {
+        wr.on("error", function(err: any) {
             reject(err);
         });
-        wr.on("close", function (ex: any) {
+        wr.on("close", function(ex: any) {
             resolve();
         });
         rd.pipe(wr);
@@ -109,12 +108,13 @@ function copyFile(source: string, target: string, ignoreNotExistent?: boolean) {
 
 class FileExtractor {
     static extractAll(files: FileMoveObject[], inputRoot: string, outputRoot: string, timeout?: number) {
-        var promises : Promise<void>[] = [];
+        var promises: Promise<void>[] = [];
 
         for(var file of files) {
+            if(!file.inputPath || !file.relativeOutputPath) continue;
             var inputPath = path.isAbsolute(file.inputPath) ?
-                                path.resolve(file.inputPath).replace(/\\/g, "/")
-                                : path.resolve(`${inputRoot}/${file.inputPath}`).replace(/\\/g, "/");
+                path.resolve(file.inputPath).replace(/\\/g, "/")
+                : path.resolve(`${inputRoot}/${file.inputPath}`).replace(/\\/g, "/");
             var outputPath = path.resolve(`${outputRoot}/${file.relativeOutputPath}`).replace(/\\/g, "/");
 
             var promise = copyFile(inputPath, outputPath, true);
@@ -127,7 +127,7 @@ class FileExtractor {
     }
 
     static replaceAllLinks(html: string) {
-        var files : FileMoveObject[] = [];
+        var files: FileMoveObject[] = [];
         var fileExtractorObject;
 
         // process images