+++ src/FileMoveObject.ts
@@ -4,9 +4,9 @@ class FileMoveObject {
     inputPath: string;
     relativeOutputPath: string;
 
-    constructor(inputPath?: string, relativeOutputPath?: string) {
-        if(inputPath !== undefined) this.inputPath = inputPath;
-        if(relativeOutputPath !== undefined) this.relativeOutputPath = relativeOutputPath;
+    constructor(inputPath: string, relativeOutputPath: string) {
+        this.inputPath = inputPath;
+        this.relativeOutputPath = relativeOutputPath;
     }
 }